+++ ntex-bytes/src/lib.rs
@@ -69,5 +69,5 @@ mod serde;
 mod string;
 
 pub use crate::bytes::{Bytes, BytesMut};
-pub use crate::pool::Pool;
+pub use crate::pool::{Pool, PoolId};
 pub use crate::string::ByteString;